load: split git-revision loading into git.go; add withCapture helper
Move the eight git-revision functions (loadFromGitRevision, readGitRefContent,
hintMissingObject, refBeforeColon, gitFetch, commitExistsLocally, blobHashFromRef,
gitShow) out of load.go into git.go, leaving load.go with just the from dispatcher
and the URL helpers. Also extract withCapture so from's URL and file branches stop
repeating the copy-loader-and-install-recorder dance; behavior unchanged.
